Refactor: Split Modal Component
