+++ bridge-project/src/page/FamPage/CreateForm.tsx
@@ -2,10 +2,7 @@ import React, { useState } from 'react';
 import { Info } from '../../model/info';
 import { TabModel } from '../../model/tabModel';
 import { addPrayerRequest } from '../../lib/firestore/card';
-import { createUser } from '../../lib/firestore';
-import Modal from 'react-modal';
-import { addUsertoCell } from '../../lib/firestore/cell';
-import { toast, ToastContainer } from 'react-toastify';
+import CreateUserModal from './CreateUserModal';
 
 interface CreateFormProps {
   curDate: string;
@@ -96,7 +93,7 @@ const CreateForm: React.FC<CreateFormProps> = ({ curDate, categories, changeIsWr
       if (error instanceof Error) {
         // 모달 오픈
         // 모달 내부에서, createUser
-        openModal();
+        setIsModalOpen(true);
       } else {
         console.error('Error adding prayer request:', error);
       }
@@ -108,86 +105,23 @@ const CreateForm: React.FC<CreateFormProps> = ({ curDate, categories, changeIsWr
     changeIsWriting(false);
   };
 
-  const customStyles = {
-    content: {
-      top: '50%',
-      left: '50%',
-      right: 'auto',
-      bottom: 'auto',
-      marginRight: '-50%',
-      transform: 'translate(-50%, -50%)',
-    },
-  };
   const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
-
-  const openModal = () => {
-    setIsModalOpen(true);
-  };
 
   const closeModal = () => {
     setIsModalOpen(false);
   };
 
-  const createMember = async () => {
-    try {
-      const userId = await createUser({ name: newInfo.name, level: 10 });
-      await addUsertoCell(userId, newInfo.cellId);
-      closeModal();
-      successToast();
-    } catch (error) {
-      console.log('Error while create new member: ', error);
-    }
-  };
-
-  const successToast = () =>
-    toast.success(`${newInfo.name} 이/가 셀에 추가되었습니다. 이어서 저장을 눌러주세요.`, {
-      position: 'bottom-left',
-      autoClose: 1500,
-      hideProgressBar: false,
-      closeOnClick: true,
-      pauseOnHover: true,
-      draggable: true,
-      progress: undefined,
-      theme: 'light',
-    });
-
   return (
     <div className="flex flex-col px-2 space-y-2">
-      <ToastContainer />
-      <Modal
+      <button onClick={() => setIsModalOpen(true)}></button>
+      <CreateUserModal
         isOpen={isModalOpen}
-        onRequestClose={closeModal}
-        style={customStyles}
-        contentLabel="Example Modal"
-        ariaHideApp={false}
-      >
-        <div className="container">
-          <p className="text-lg text-center">셀원 정보 확인</p>
-          <div className="py-6">
-            현재 팸에는 "{newInfo.name}" 이름의 셀원 정보가 없습니다. <br />
-            <select
-              className="w-1/3 input-box"
-              value={newInfo.cellId}
-              onChange={(e) => createCategory(e.target.value)}
-            >
-              {categories.map((item) => (
-                <option value={item.id} key={item.id}>
-                  {item.name}
-                </option>
-              ))}
-            </select>
-            의 새로운 셀원으로 추가하시겠습니까?
-          </div>
-          <div className="flex flex-row justify-end space-x-2">
-            <button onClick={closeModal} type="button" className="self-end outline-hover-btn">
-              수정하기
-            </button>
-            <button onClick={createMember} type="button" className="self-end primary-hover-btn">
-              추가하기
-            </button>
-          </div>
-        </div>
-      </Modal>
+        cellId={newInfo.cellId}
+        cellCategory={categories}
+        updateCategoty={createCategory}
+        userName={newInfo.name}
+        closeModal={closeModal}
+      />
       <div className="flex space-x-2">
         <input
           type="text"
